ref: move fetching function to i18n module
