--- packages/cli/src/utils/doesLocaleFileExists.ts
@@ -1,36 +1,0 @@
-/*
- * Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { existsSync } from 'fs-extra';
-
-const doesLocaleFileExists = (locale: string) => {
-  const localeArray = [
-    locale,
-    locale.split('-')[0],
-    locale.split('_')[0],
-    'en',
-  ];
-
-  for (const _locale of localeArray) {
-    if (existsSync(`_locales/${_locale}/messages.json`)) {
-      return _locale;
-    }
-  }
-
-  return 'en';
-};
-
-export default doesLocaleFileExists;
